+++ packages/androidx_graphics_shapes/lib/src/material_shapes/material_shapes_compose.dart
@@ -1,0 +1,1 @@
+export 'material_shapes.dart';